os/bluestore: Move Blob to bluestore namespace in BlueStore_objects
(cherry picked from commit feb748b34c62e11daeeaeb409cef510b12464a1a)
(cherry picked from commit 06c417e8b4a81fe445f93ae668f1e8c0adca199b)
