+++ src/os/bluestore/BlueStore.h
@@ -68,6 +68,11 @@ class Allocator;
 class FreelistManager;
 class BlueStoreRepairer;
 class SimpleBitmap;
+
+namespace bluestore {
+  struct Blob;
+}
+
 //#define DEBUG_CACHE
 //#define DEBUG_DEFERRED
 
@@ -638,180 +643,8 @@ public:
     void dump(CephContext *cct);
   };
 
-  /// in-memory blob metadata and associated cached buffers (if any)
-  struct Blob {
-    MEMPOOL_CLASS_HELPERS();
+  using Blob = bluestore::Blob;
 
-    std::atomic_int nref = {0};     ///< reference count
-    int16_t id = -1;                ///< id, for spanning blobs only, >= 0
-    int16_t last_encoded_id = -1;   ///< (ephemeral) used during encoding only
-    CollectionRef collection;
-
-    void set_shared_blob(SharedBlobRef sb) {
-      ceph_assert((bool)sb);
-      ceph_assert(!shared_blob);
-      ceph_assert(sb->collection = collection);
-      shared_blob = sb;
-      ceph_assert(get_cache());
-    }
-    Blob(CollectionRef collection) : collection(collection) {}
-  private:
-    SharedBlobRef shared_blob;      ///< shared blob state (if any)
-    mutable bluestore_blob_t blob;  ///< decoded blob metadata
-    /// refs from this shard.  ephemeral if id<0, persisted if spanning.
-    bluestore_blob_use_tracker_t used_in_blob;
-
-  public:
-
-    friend void intrusive_ptr_add_ref(Blob *b) { b->get(); }
-    friend void intrusive_ptr_release(Blob *b) { b->put(); }
-
-    void dump(ceph::Formatter* f) const;
-    friend std::ostream& operator<<(std::ostream& out, const Blob &b);
-    struct printer : public BlueStore::printer {
-      const Blob& blob;
-      uint16_t mode;
-      printer(const Blob& blob, uint16_t mode)
-      :blob(blob), mode(mode) {}
-    };
-    friend std::ostream& operator<<(std::ostream& out, const printer &p);
-    printer print(uint16_t mode) const {
-      return printer(*this, mode);
-    }
-    const bluestore_blob_use_tracker_t& get_blob_use_tracker() const {
-      return used_in_blob;
-    }
-    bluestore_blob_use_tracker_t& dirty_blob_use_tracker() {
-      return used_in_blob;
-    }
-
-    const SharedBlobRef& get_shared_blob() const {
-      return shared_blob;
-    }
-
-    SharedBlobRef& get_dirty_shared_blob() {
-      return shared_blob;
-    }
-
-    bool is_referenced() const {
-      return used_in_blob.is_not_empty();
-    }
-    uint32_t get_referenced_bytes() const {
-      return used_in_blob.get_referenced_bytes();
-    }
-
-    bool is_spanning() const {
-      return id >= 0;
-    }
-
-    bool can_split() {
-      // splitting a BufferSpace writing list is too hard; don't try.
-      return used_in_blob.can_split() &&
-             get_blob().can_split();
-    }
-
-    bool can_merge_blob(const Blob* other, uint32_t& blob_end) const;
-    uint32_t merge_blob(CephContext* cct, Blob* blob_to_dissolve);
-
-    bool can_split_at(uint32_t blob_offset) const {
-      return used_in_blob.can_split_at(blob_offset) &&
-             get_blob().can_split_at(blob_offset);
-    }
-
-    bool can_reuse_blob(uint32_t min_alloc_size,
-			uint32_t target_blob_size,
-			uint32_t b_offset,
-			uint32_t *length0);
-
-    void dup(Blob& o) {
-      o.set_shared_blob(shared_blob);
-      o.blob = blob;
-    }
-    void add_tail(uint32_t new_blob_size, uint32_t min_release_size);
-    void dup(const Blob& from, bool copy_used_in_blob);
-    void copy_from(CephContext* cct, const Blob& from,
-		   uint32_t min_release_size, uint32_t start, uint32_t len);
-    void copy_extents(CephContext* cct, const Blob& from, uint32_t start,
-		      uint32_t pre_len, uint32_t main_len, uint32_t post_len);
-    void copy_extents_over_empty(CephContext* cct, const Blob& from, uint32_t start, uint32_t len);
-
-    inline const bluestore_blob_t& get_blob() const {
-      return blob;
-    }
-    inline bluestore_blob_t& dirty_blob() {
-      return blob;
-    }
-
-    /// get logical references
-    void get_ref(Collection *coll, uint32_t offset, uint32_t length);
-    /// put logical references, and get back any released extents
-    bool put_ref(Collection *coll, uint32_t offset, uint32_t length,
-		 PExtentVector *r);
-    uint32_t put_ref_accumulate(
-      Collection *coll,
-      uint32_t offset,
-      uint32_t length,
-      PExtentVector *released_disk);
-    /// split the blob
-    void split(Collection *coll, uint32_t blob_offset, Blob *o);
-
-    void maybe_prune_tail();
-
-    void get() {
-      ++nref;
-    }
-    void put() {
-      if (--nref == 0)
-	delete this;
-    }
-    bool is_shared_loaded() const {
-      return shared_blob && shared_blob->is_loaded();
-    }
-    inline BufferCacheShard* get_cache() {
-      return collection ? collection->cache : nullptr;
-    }
-    uint64_t get_sbid() const {
-      return shared_blob ? shared_blob->get_sbid() : 0;
-    }
-    CollectionRef get_collection() const {
-      return collection;
-    }
-
-    ~Blob();
-
-    void bound_encode(
-      size_t& p,
-      uint64_t struct_v,
-      uint64_t sbid,
-      bool include_ref_map) const {
-      denc(blob, p, struct_v);
-      if (blob.is_shared()) {
-        denc(sbid, p);
-      }
-      if (include_ref_map) {
-	used_in_blob.bound_encode(p);
-      }
-    }
-    void encode(
-      ceph::buffer::list::contiguous_appender& p,
-      uint64_t struct_v,
-      uint64_t sbid,
-      bool include_ref_map) const {
-      denc(blob, p, struct_v);
-      if (blob.is_shared()) {
-        denc(sbid, p);
-      }
-      if (include_ref_map) {
-	used_in_blob.encode(p);
-      }
-    }
-    void decode(
-      ceph::buffer::ptr::const_iterator& p,
-      uint64_t struct_v,
-      uint64_t* sbid,
-      bool include_ref_map,
-      Collection *coll);
-  };
   typedef boost::intrusive_ptr<Blob> BlobRef;
   typedef mempool::bluestore_cache_meta::map<int,BlobRef> blob_map_t;
 
@@ -836,11 +669,7 @@ public:
         logical_offset(lo), blob_offset(o), length(l) {
       assign_blob(b);
     }
-    ~Extent() {
-      if (blob) {
-	blob->get_cache()->rm_extent();
-      }
-    }
+    ~Extent();
     struct printer : public BlueStore::printer {
       const Extent& ext;
       uint16_t mode;
@@ -854,11 +683,7 @@ public:
 
     void dump(ceph::Formatter* f) const;
 
-    void assign_blob(const BlobRef& b) {
-      ceph_assert(!blob);
-      blob = b;
-      blob->get_cache()->add_extent();
-    }
+    void assign_blob(const BlobRef& b);
 
     // comparators for intrusive_set
     friend bool operator<(const Extent &a, const Extent &b) {
@@ -875,9 +700,7 @@ public:
       return logical_offset - blob_offset;
     }
 
-    uint32_t blob_end() const {
-      return blob_start() + blob->get_blob().get_logical_length();
-    }
+    uint32_t blob_end() const;
 
     uint32_t logical_end() const {
       return logical_offset + length;
@@ -1058,7 +881,7 @@ public:
       KeyValueDB::Transaction t,
       bool just_after_reshard //true to indicate that update should now respect shard boundaries
     );                        //as no further resharding will be done
-    decltype(BlueStore::Blob::id) allocate_spanning_blob_id();
+    int16_t allocate_spanning_blob_id();
     void reshard(
       KeyValueDB *db,
       KeyValueDB::Transaction t,
@@ -1672,11 +1495,7 @@ public:
     void make_blob_shared(uint64_t sbid, BlobRef b);
     uint64_t make_blob_unshared(SharedBlob *sb);
 
-    BlobRef new_blob() {
-      BlobRef b = new Blob(this);
-      b->get_cache()->add_blob();
-      return b;
-    }
+    BlobRef new_blob();
 
     bool contains(const ghobject_t& oid) {
       if (cid.is_meta())
